Clean up and refactor
